requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs
CacheControl==0.14.4
certifi
cffi==2.0.0
charset-normalizer
cloudinary
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
Django==5.2.12
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
firebase_admin
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-firestore
google-cloud-storage
google-crc32c==1.8.0
google-genai
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.78.1
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperframe==6.1.0
idna
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
msgpack==1.1.2
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pyparsing==3.3.2
python-dotenv
PyYAML==6.0.3
referencing==0.37.0
requests    
rpds-py==0.30.0
rsa==4.9.1
six
sniffio
sqlparse==0.5.5
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions
tzdata==2025.3
uritemplate==4.2.0
urllib3
websockets==16.0